[explorer/puller] task: change data type

+++ explorer/puller/db/NemDatabase.py
@@ -8,10 +8,12 @@ class NemDatabase(DatabaseConnection):
 		"""Creates blocks database tables."""
 
 		cursor = self.connection.cursor()
+
+		# Create blocks table
 		cursor.execute('''CREATE TABLE IF NOT EXISTS blocks (
-			height integer NOT NULL,
+			height bigint NOT NULL,
 			timestamp timestamp NOT NULL,
-			totalFees integer DEFAULT 0,
+			totalFees bigint DEFAULT 0,
 			totalTransactions integer DEFAULT 0,
 			difficulty bigInt NOT NULL,
 			hash VARCHAR(64) NOT NULL,
